back_end/pyproject.toml
[tool.poetry]
name = "back_end"
version = "0.1.0"
description = "Soul School back_end"
authors = ["Jojoah Kaseq <egyabig2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
djangorestframework = "^3.14.0"
djoser = "^2.2.2"
drf-yasg = "^1.21.7"
django-behaviors = "^0.5.1"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
django-environ = "^0.11.2"
django-cors-headers = "^4.3.1"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.3.0"
pytest-django = "^4.8.0"
psycopg2 = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
